+++ plugins/ipam/host-local/backend/disk/lock.go
@@ -1,3 +1,4 @@
+// +build !windows
 // Copyright 2015 CNI authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
